nice code refactor, over 50% good answers for UC3, no kafka deploy yet
