+++ src/main/java/allezon/config/KafkaStreamsConfig.java
@@ -6,32 +6,22 @@ import org.apache.kafka.streams.StreamsBuilder;
 import org.apache.kafka.streams.StreamsConfig;
 import org.apache.kafka.streams.KeyValue;
 import org.apache.kafka.streams.kstream.*;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.kafka.support.serializer.JsonSerde;
 
 import com.aerospike.client.AerospikeClient;
-import com.aerospike.client.AerospikeException;
-import com.aerospike.client.Bin;
-import com.aerospike.client.Key;
-import com.aerospike.client.policy.GenerationPolicy;
-import com.aerospike.client.policy.RecordExistsAction;
-import com.aerospike.client.policy.WritePolicy;
 
 import allezon.domain.AggregatedValue;
 import allezon.domain.UserTagEvent;
-import allezon.service.UserActionsService;
+import allezon.service.AerospikeService;
 
-import static allezon.constant.Constants.COUNT_BIN;
-import static allezon.constant.Constants.NAMESPACE;
-import static allezon.constant.Constants.PRICE_BIN;
-import static allezon.constant.Constants.SET_ANALYTICS;
+import static allezon.constant.Constants.ANALYTICS_INPUT_TOPIC;
+import static allezon.constant.Constants.APP_ID;
 import static allezon.constant.Constants.BLANK;
 import static allezon.constant.Constants.DELIMITER;
-import com.aerospike.client.Record;
-import com.aerospike.client.ResultCode;
 
 import java.time.Duration;
 import java.time.Instant;
@@ -43,22 +33,27 @@ import java.util.List;
 import java.util.Properties;
 
 @Configuration
-public class KafkaStreamsConfig {
-    private static final Logger log = LoggerFactory.getLogger(KafkaStreamsConfig.class);
-    private final static String APP_ID = "allezonappka";
-    public final static String STATE_STORE_NAME_KEY_VALUE_NAME = "allss";
-    public final static String ANALYTICS_INPUT_TOPIC = "analyti";                     
+public class KafkaStreamsConfig {                    
     public static final DateTimeFormatter BUCKET_FORMATTER = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss").withZone(ZoneOffset.UTC);
 
+    @Value("${spring.kafka.streams.bootstrap-servers}")
+    private String bootstrapServers;
+
+    private final AerospikeService aerospikeService;
+
+    @Autowired
+    public KafkaStreamsConfig(AerospikeService aerospikeService) {
+        this.aerospikeService = aerospikeService;
+    }
 
     @Bean
     public Properties kafkaStreamsProperties() {
         Properties props = new Properties();
         props.put(StreamsConfig.APPLICATION_ID_CONFIG, APP_ID);
-        props.put(StreamsConfig.BOOTSTRAP_SERVERS_CONFIG, "st124vm101.rtb-lab.pl:9092,st124vm102.rtb-lab.pl:9092,st124vm103.rtb-lab.pl:9092");
+        props.put(StreamsConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapServers);
         props.put(StreamsConfig.DEFAULT_KEY_SERDE_CLASS_CONFIG, Serdes.String().getClass().getName());
         props.put(StreamsConfig.DEFAULT_VALUE_SERDE_CLASS_CONFIG, JsonSerde.class.getName());
-        String stateDir = "/tmp/kafka-streams/" + System.getenv("HOSTNAME");
+        String stateDir = "/tmp/kafka-streams/" + System.getenv("HOSTNAME") + "allezon";
         System.out.println("set state dir to: " + stateDir);
         props.put(StreamsConfig.STATE_DIR_CONFIG, stateDir);
         props.put(StreamsConfig.NUM_STREAM_THREADS_CONFIG, 10);  // Ensures matching partitions
@@ -81,7 +76,7 @@ public class KafkaStreamsConfig {
         KTable<Windowed<String>, AggregatedValue> aggregatedTable = userTagEventsStream
             .flatMap((key, userTagEvent) -> reKeyInputStream(userTagEvent))
             .groupByKey(Grouped.with(Serdes.String(), Serdes.Integer()))
-            .windowedBy(TimeWindows.ofSizeAndGrace(Duration.ofSeconds(80), Duration.ofSeconds(10)))
+            .windowedBy(TimeWindows.ofSizeAndGrace(Duration.ofSeconds(30), Duration.ofSeconds(5)))
             .aggregate(
                 AggregatedValue::new,
                 (key, event, aggregate) -> aggregate.aggregateProduct(event),
@@ -90,51 +85,9 @@ public class KafkaStreamsConfig {
         .suppress(Suppressed.untilWindowCloses(Suppressed.BufferConfig.unbounded())); // Suppress until the window closes
 
         aggregatedTable.toStream().foreach((key, value) -> {
-            while (true) {
-                try {
-                    Key aerospikeKey = new Key(NAMESPACE, SET_ANALYTICS, key.key());
-    
-                    // Fetch existing record if available
-                    Record existingRecord = client.get(null, aerospikeKey);
-                
-                    // Initialize existing values as 0 if the record doesn't exist
-                    int existingCount = 0;
-                    int existingPrice = 0;
-                
-                    if (existingRecord != null) {
-                        existingCount = existingRecord.getInt(COUNT_BIN); // Get the existing count
-                        existingPrice = existingRecord.getInt(PRICE_BIN); // Get the existing price
-                    }
-                
-                    // Add the new values to the existing ones
-                    int newCount = existingCount + value.getCount();
-                    int newPrice = existingPrice + value.getPrice();
-                
-                    // Create new bins with the accumulated values
-                    Bin countBin = new Bin(COUNT_BIN, newCount);
-                    Bin priceBin = new Bin(PRICE_BIN, newPrice);
-                
-                    WritePolicy writePolicy = new WritePolicy();
-                    writePolicy.recordExistsAction = RecordExistsAction.UPDATE;  
-                    writePolicy.generationPolicy = (existingRecord != null) ? GenerationPolicy.EXPECT_GEN_EQUAL : GenerationPolicy.NONE;
-                    writePolicy.generation = (existingRecord != null) ? existingRecord.generation : 0;
-                    writePolicy.expiration = 24 * 60 * 60;
-                
-                    // Put the accumulated values back into Aerospike
-                    client.put(writePolicy, aerospikeKey, countBin, priceBin);
-                    break;
-                } catch (AerospikeException ae) {
-                    if (ae.getResultCode() == ResultCode.GENERATION_ERROR) {
-                        log.info("Generation error, retrying...");
-                        continue;
-                    } else {
-                        log.error("Error in addEvent", ae);
-                        throw ae; // For other exceptions, rethrow
-                    }
-                } catch (Exception e) {
-                    log.error("Some other exception", e);
-                    throw e;
-                }
+            boolean success = false;
+            while (!success) {
+                success = aerospikeService.writeAggregatesToAerospike(key.key(), value);
             }
         });
