+++ src/main/resources/application.properties
@@ -2,7 +2,7 @@ aerospike.seeds=st124vm102.rtb-lab.pl,st124vm103.rtb-lab.pl,st124vm104.rtb-lab.p
 aerospike.port=3000
 server.port=8085
 
-spring.kafka.streams.application-id=allezonappka
+spring.kafka.streams.application-id=app-allezon
 spring.kafka.streams.bootstrap-servers=st124vm101.rtb-lab.pl:9092,st124vm102.rtb-lab.pl:9092,st124vm103.rtb-lab.pl:9092
 spring.kafka.streams.default-key-serde-class=org.apache.kafka.common.serialization.Serdes$StringSerde
 spring.kafka.streams.default-value-serde-class=org.springframework.kafka.support.serializer.JsonSerde
